+++ src/components/index.ts
@@ -3,10 +3,9 @@ export * from './card-item';
 export * from './category-button';
 export * from './category-selector';
 export * from './container';
-export * from './container';
 export * from './counter';
 export * from './icon';
-export * from './icon';
 export * from './input';
+export * from './list-card-menu';
 export * from './navbar';
 export * from './search-input';
