+++ src/data/data-menu.ts
@@ -12,6 +12,17 @@ export type MenuItem = {
   additionalIds: AdditionId[];
 };
 
+export interface Category {
+  id: string;
+  name: string;
+  subCategories: SubCategory[];
+}
+
+export interface SubCategory {
+  id: string;
+  name: string;
+}
+
 export const categories = [
   {
     id: 'entradas',
